+++ src/player/resources/character.rs
@@ -35,15 +35,15 @@ pub struct Character {
     /// Period of time between firing blasts
     pub fire_period: f32,
     /// Health of the player
-    pub health: f32,
+    pub health: usize,
     /// Shields of the player
-    pub shields: f32,
+    pub shields: usize,
     /// Shields recharging rate
     pub shields_recharge_rate: f32,
     /// Amount of damage dealt per attack
-    pub attack_damage: f32,
+    pub attack_damage: usize,
     /// Amount of damage dealt on contact
-    pub collision_damage: f32,
+    pub collision_damage: usize,
     /// Distance to attract items and consumables
     pub attraction_distance: f32,
     /// Acceleration applied to items and conumables in attraction distance
