+++ src/player/systems/ability.rs
@@ -199,7 +199,7 @@ pub fn player_ability_system(
                     spawn_projectile.send(SpawnProjectileEvent {
                         projectile_type: player_component.projectile_type.clone(),
                         transform: projectile_transform,
-                        damage: player_component.attack_damage * multiplier,
+                        damage: (player_component.attack_damage as f32 * multiplier) as usize,
                         despawn_time: player_component.projectile_despawn_time,
                         initial_motion,
                         source: entity,
